ENH: Add caching of scraped content
Caching of content scraped from websites has been added to help avoid
servers blocking the scraper due to too many requests. The cache is
a dictionary with keys corresponding to the exact URLs scraped and
values containing the content of their pages. The cache itself is
stored as a pickle file in the data directory.

Partially addresses #3